Extract macro tests to separate package
