+++ Cargo.toml
@@ -18,6 +18,7 @@ members = [
   "crates/turbo-tasks-hash",
   "crates/turbo-tasks-macros",
   "crates/turbo-tasks-macros-shared",
+  "crates/turbo-tasks-macros-tests",
   "crates/turbo-tasks-memory",
   "crates/turbo-tasks-testing",
   "crates/turbo-updater",
@@ -67,6 +68,7 @@ default-members = [
   "crates/turbo-tasks-hash",
   "crates/turbo-tasks-macros",
   "crates/turbo-tasks-macros-shared",
+  "crates/turbo-tasks-macros-tests",
   "crates/turbo-tasks-memory",
   "crates/turbo-tasks-testing",
   "crates/turbopack",
@@ -123,6 +125,7 @@ turbo-tasks-fs = { path = "crates/turbo-tasks-fs" }
 turbo-tasks-hash = { path = "crates/turbo-tasks-hash" }
 turbo-tasks-macros = { path = "crates/turbo-tasks-macros" }
 turbo-tasks-macros-shared = { path = "crates/turbo-tasks-macros-shared" }
+turbo-tasks-macros-tests = { path = "crates/turbo-tasks-macros-tests" }
 turbo-tasks-memory = { path = "crates/turbo-tasks-memory" }
 turbo-tasks-testing = { path = "crates/turbo-tasks-testing" }
 turbo-updater = { path = "crates/turbo-updater" }
