+++ crates/turbo-tasks-macros-tests/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "turbo-tasks-macros-tests"
+version = "0.1.0"
+description = "TBD"
+license = "MPL-2.0"
+edition = "2021"
+
+[dev-dependencies]
+anyhow = { workspace = true }
+tokio = { workspace = true }
+turbo-tasks = { workspace = true }
+turbo-tasks-memory = { workspace = true }
+turbo-tasks-testing = { workspace = true }
+# TODO: turbo-tasks-testing uses a macro that requires us to depend on lazy-static.
+lazy_static = { workspace = true }
+
+[build-dependencies]
+turbo-tasks-build = { workspace = true }
