+++ crates/turbo-tasks-macros/Cargo.toml
@@ -17,14 +17,3 @@ proc-macro2 = { workspace = true }
 quote = { workspace = true }
 syn = { workspace = true, features = ["full", "extra-traits"] }
 turbo-tasks-macros-shared = { workspace = true }
-
-[dev-dependencies]
-# TODO: turbo-tasks-testing uses a macro that requires us to depend on lazy-static.
-lazy_static = { workspace = true }
-tokio = { workspace = true }
-turbo-tasks = { workspace = true }
-turbo-tasks-memory = { workspace = true }
-turbo-tasks-testing = { workspace = true }
-
-[build-dependencies]
-turbo-tasks-build = { workspace = true }
